pyproject.toml
[project]
name = "LLM-Morphseg-Research"
version = "0.1.0"
description = "Research project exploring the use of large language models for morphological segmentation"
requires-python = ">=3.13"

dependencies = [
    "bitsandbytes==0.49.2",
    "datasets==4.5.0",
    "peft==0.18.1",
    "torch==2.10.0",
    "transformers==5.2.0",
    
    "accelerate==1.12.0",
    "numpy==2.4.2",
    "scipy==1.17.0",
    "pandas<3.0.0",
    "huggingface-hub==1.4.1",
    
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    
    "mlflow>=3.10.0",
    
    "pytorch-lightning>=2.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/morphseg"]

[project.scripts]
do-train = "morphseg.cli:train"
run-inference = "morphseg.cli:inference"

